Refactor RemoteApiService with request and response helpers
